[AviSynthAudioEncoder] force the 2nd portion to be long. Patch by ateeq.
git-svn-id: http://svn.code.sf.net/p/megui/code@748 a21b8521-4df1-4ed3-b48c-cff5bc73dcb3

+++ megui/trunk/AviSynthAudioEncoder.cs
@@ -474,7 +474,7 @@ new JobProcessorFactory(new ProcessorFactory(init), "AviSynthAudioEncoder");
         {
             const uint FAAD_MAGIC_VALUE = 0xFFFFFF00;
             const uint WAV_HEADER_SIZE = 36;
-            bool useFaadTrick = a.AudioSizeInBytes >= (uint.MaxValue - WAV_HEADER_SIZE);
+            bool useFaadTrick = a.AudioSizeInBytes >= ((long)uint.MaxValue - WAV_HEADER_SIZE);
             target.Write(System.Text.Encoding.ASCII.GetBytes("RIFF"), 0, 4);
             target.Write(BitConverter.GetBytes(useFaadTrick ? FAAD_MAGIC_VALUE : (uint)(a.AudioSizeInBytes + WAV_HEADER_SIZE)), 0, 4);
             target.Write(System.Text.Encoding.ASCII.GetBytes("WAVEfmt "), 0, 8);
